stream: move legacy to sep file
