+++ node.gyp
@@ -61,6 +61,7 @@
       'lib/_stream_transform.js',
       'lib/_stream_passthrough.js',
       'lib/_stream_wrap.js',
+      'lib/_stream_legacy.js',
       'lib/string_decoder.js',
       'lib/sys.js',
       'lib/timers.js',
